update int to long

+++ src/main/java/com/meli/football_matches_api/controller/MatchController.java
@@ -31,7 +31,7 @@ public class MatchController {
     }
 
     @DeleteMapping(params = "id")
-    public ResponseEntity<String> delete(@RequestParam int id) {
+    public ResponseEntity<String> delete(@RequestParam Long id) {
         return matchService.delete(id);
     }
